pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "g-stvk-flow"
version = "0.1.0"
description = "Geometric Spatio-Temporal Variance-preserving K-Flow"
readme = "README.md"
requires-python = ">=3.10"
authors = [{name = "G-STVK-Flow Engineering"}]
dependencies = [
  "torch>=2.2.0",
  "torchvision>=0.17.0",
  "PyYAML>=6.0",
  "numpy>=1.24.0",
  "tqdm>=4.66.0",
]

[tool.setuptools.packages.find]
include = ["g_stvk_flow*"]
